+++ src/main/java/net/proselyte/springbootdemo/controller/OrderController.java
@@ -34,17 +34,28 @@ public class OrderController {
   @GetMapping("/get/{id}")
   public ResponseEntity<?> getOrderById(@PathVariable Long id) {
     Order order = orderService.findById(id);
+    if (order == null) {
+      return ResponseEntity.notFound().build();
+    }
     return new ResponseEntity<>(order, HttpStatus.OK);
   }
 
   @PutMapping("/update/{id}")
   public ResponseEntity<?> update(@PathVariable Long id, @RequestBody OrderDto orderDto) {
+    Order order = orderService.findById(id);
+    if (order == null) {
+      return ResponseEntity.notFound().build();
+    }
     Order updatedOrder = orderService.updateOrder(id, orderDto);
     return new ResponseEntity<>(updatedOrder, HttpStatus.OK);
   }
 
   @DeleteMapping("/delete/{id}")
   public ResponseEntity<?> delete(@PathVariable Long id) {
+   Order order = orderService.findById(id);
+    if (order == null) {
+      return ResponseEntity.notFound().build();
+    }
     orderService.deleteById(id);
     return ResponseEntity.status(HttpStatus.OK).build();
   }
